LF-9610 feat(imx8ulp): add imx8ulp apd dsl support
Add i.MX8ULP DSL mode support. this mode can NOT be
enabled if the PD mode is enabled. So add a build
macro for it. if anyone want to use this mode, need to
rebuild the BL31 with 'IMX8ULP_DSL_SUPPORT=1'.

for normal PSCI system suspend mode, the CPU core should
enter power down status. but for i.MX8ULP DSL mode, the
cpu core is just clock gated during the DSL mode, so need
to modify the PSCI common code to enter 'PSTATE_TYPE_STANDBY'
mode

Signed-off-by: Jacky Bai <ping.bai@nxp.com>
Reviewed-by: Ye Li <ye.li@nxp.com>
Change-Id: I95604a575610714b078ce1a249e22e9d1ab3dccf