LF-13319 feat(imx95): Implement a semaphore for GIC quiescing between SM and ATF
As there is no way to reset the GIC on iMX95, SM needs to
quiesce the GIC using the GIC waker. But AP can also be touching
GIC waker register during entry/exit from low power modes.
Add a semaphore to provide a critical section for GIC waker.
The last two words in the ATF MU 1K SRAM space is used to hold
the semaphore (follow the Peterson's algorithm for mutual exclusion).

Signed-off-by: Ranjani Vaidyanathan <ranjani.vaidyanathan@nxp.com>