MA-23254 [8QXP] Security configure for widevine L1 on imx8qxp
Configure normal memory for vpu_part to FULL:
When decoding starts, the VPU needs to be moved to the VPU parition, so
the VPU needs to have access to the memory region in the os part.

Add smcc monitor command to handle that.

Allocate DPU parition:
The attribute of dpu part is grant=true, Isolated=true
grant is true: linux can control, and secure world can
               set misc control for dpu
Isolated=true: assign new domain id and can't w/r tee memory.

Add monitor smccall to get the number of dpu/os part:
Because trusty can't get the dpu part and os part number,
So there need to add one fastcall to get it, and return it
to Trusty.

Change-Id: I0d9d62e4cf62c2cc09519d49d4b199b08fbe8e2b
Signed-off-by: He Zhai <zhai.he@nxp.com>