LF-14639 fix(imx94): stop m33s once before start
The netcmix has reset and power handshake events to ELE. For the first time
netcmix power on, it will trigger netc reset event and m33s reset event to
ELE, but if the netcmix power off and on after the first time power on, only
power event is triggered to ELE, no m33s reset handshake event is triggered
to ELE, so ELE will not release the m33s core in the netcmix power on sequence.
m33s can't be started successfully if ELE does not release the m33s core.

To make sure the m33s is release by ELE before m33s start, we need to call stop
m33s to let SM to do a complete reset cycle to trigger the reset event to ELE to
release m33s.

Signed-off-by: Jacky Bai <ping.bai@nxp.com>