LF-14579 fix(imx8ulp): retry if upower return UPWR_RESP_RESOURCE err
When the upower is in the critical section(protected by mutex) to
handle hotplug event, while A35 core0 is calling the upower API
to power up the power domain, the API call can not take the mutex,
upower will return 'UPWR_RESP_RESOURCE' err to the caller, so the
power domain power up failed. We need to redo the upower API call
if the err id is 'UPWR_RESP_RESOURCE'.

Signed-off-by: Jacky Bai <ping.bai@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>