LF-15212 fix(imx8m): add mask for jr0mid check
In current code, if the CAAM_JR0MID is 0x8011, that means this Job Ring is
reserved for HAB purpose. But In some conditions, the value read from the
CAAM_JR0MID is 0x80008011, the LDID(bit31) will also be set to lock the
register. In such case, th JR0 is also reserved for HAB. So use a value
mask to only use the lower 16 bits to check if the CAAM JR0 is owned
by secure world.

Signed-off-by: Jacky Bai <ping.bai@nxp.com>
Reviewed-by: Ye Li <ye.li@nxp.com>