LF-15882 fix(imx93): update the refresh rate setting when dyn_ref disabled
The DDRC DYN_REF feature performs periodic (autonomous) reads of the LP4
MR4 and automatically adjusts the auto refresh rate based on the MR4
returned values. When disabling DYN_REF, the DDRC reverts the auto refresh
rate back to 1x nominal.  There may be cases where the operating
temperature of the DRAM is high enough such that the DRAM will require a
higher refresh rate (a reduced refresh rate interval of 0.5x or 0.25x
refresh).

To ensure that the DDRC is providing enough refreshes at elevated
temperatures when DYN_REF is disabled, need to adjust the refresh rate
when DYN_REF is disabled.

Signed-off-by: Jacky Bai <ping.bai@nxp.com>
Reviewed-by: Ye Li <ye.li@nxp.com>